requirements.txt
pyswitchbot
